+++ basicdata/del/delbranch.php
@@ -1,6 +1,10 @@
-﻿<? include('../../change.php'); ?>
+﻿<? session_start(); ?>
+<? include('../../change.php'); ?>
 <?
+	if(!isset($_SESSION['right']) || $_SESSION['right']!='1') { exit; }
+	if(!isset($id) || trim($id)==="") { exit; }
 	include('../../connectdatabase.php');
+	$id = (int)$id;
 	$sql = "delete from branch where id_branch = '$id' ";
 	mysqli_query($connect, $sql);
 	//echo "Delete Success.";
