+++ user/deluser.php
@@ -1,8 +1,11 @@
-﻿<? include('../change.php'); ?>
+﻿<? session_start(); ?>
+<? include('../change.php'); ?>
 <?
+	if(!isset($_SESSION['right']) || $_SESSION['right']!='1') { exit; }
+	if(!isset($id) || trim($id)==="") { exit; }
 	include('../connectdatabase.php');
+	$id = (int)$id;
 	$sql = "delete from user where id_user = '$id' ";
 	mysqli_query($connect, $sql);
-	//echo "Delete Success.";
 	mysqli_close($connect);
 ?>
